2007-11-13	Till Straumann <strauman@slac.stanford.edu>
	* shared/include/byteorder.h: fixed wrong pointer-type
	of ld_le32() (uint16_t* -> uint32_t*).

+++ c/src/lib/libcpu/powerpc/shared/include/byteorder.h
@@ -34,7 +34,7 @@ extern __inline__ void st_le16(volatile uint16_t *addr, unsigned val)
 	__asm__ __volatile__ ("sthbrx %1,0,%2" : "=m" (*addr) : "r" (val), "r" (addr));
 }
 
-extern __inline__ unsigned ld_le32(volatile uint16_t *addr)
+extern __inline__ unsigned ld_le32(volatile uint32_t *addr)
 {
 	unsigned val;
